+++ sipx/__main__.py
@@ -1,5 +1,0 @@
-from sipx.cli.main import main
-
-
-if __name__ == "__main__":
-    raise SystemExit(main())