+++ sipx/backends/__init__.py
@@ -1,17 +1,3 @@
-from sipx.backends.asterisk import (
-    AsteriskAriClient,
-    AsteriskAriConfig,
-    AsteriskAriError,
-    AsteriskAriEvent,
-    AsteriskAriHttpResponse,
-    AsteriskBackend,
-    AsteriskBridge,
-    AsteriskChannel,
-    AsteriskMediaPath,
-    AsteriskMediaPortConfig,
-    AsteriskPlayback,
-    AsteriskWebSocketMediaPort,
-)
 from sipx.backends.mock import MockBackend
 from sipx.backends.native import (
     NativeSipBackend,
@@ -26,18 +12,6 @@ from sipx.backends.native import (
 )
 
 __all__ = [
-    "AsteriskAriClient",
-    "AsteriskAriConfig",
-    "AsteriskAriError",
-    "AsteriskAriEvent",
-    "AsteriskAriHttpResponse",
-    "AsteriskBackend",
-    "AsteriskBridge",
-    "AsteriskChannel",
-    "AsteriskMediaPath",
-    "AsteriskMediaPortConfig",
-    "AsteriskPlayback",
-    "AsteriskWebSocketMediaPort",
     "MockBackend",
     "NativeSipBackend",
     "NativeSipCall",